+++ apps/ledger-live-desktop/src/renderer/families/tezos/DelegateFlowModal/steps/StepValidator.tsx
@@ -3,7 +3,8 @@ import invariant from "invariant";
 import styled from "styled-components";
 import { Trans } from "react-i18next";
 import { getAccountBridge } from "@ledgerhq/live-common/bridge/index";
-import { useBakers, Baker } from "@ledgerhq/live-common/families/tezos/bakers";
+import { useBakers } from "@ledgerhq/live-common/families/tezos/bakers";
+import { Baker } from "@ledgerhq/live-common/families/tezos/types";
 import bakersWhitelistDefault from "@ledgerhq/live-common/families/tezos/bakers.whitelist-default";
 import { openURL } from "~/renderer/linking";
 import TrackPage from "~/renderer/analytics/TrackPage";
