+++ apps/ledger-live-mobile/src/families/tezos/BakerImage.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Image } from "react-native";
-import type { Baker } from "@ledgerhq/live-common/families/tezos/bakers";
+import type { Baker } from "@ledgerhq/live-common/families/tezos/types";
 import Circle from "../../components/Circle";
 
 type Props = {
