+++ libs/ledger-live-common/src/families/tezos/bakers.ts
@@ -1,59 +1,11 @@
 import { makeLRUCache } from "@ledgerhq/live-network/cache";
 import network from "@ledgerhq/live-network/network";
 import { log } from "@ledgerhq/logs";
-import type { AccountLike, Operation } from "@ledgerhq/types-live";
+import type { AccountLike } from "@ledgerhq/types-live";
 import { useEffect, useMemo, useState } from "react";
 import { getEnv } from "@ledgerhq/live-env";
 import { ledgerValidatorAddress } from "./bakers.whitelist-default";
-
-type CapacityStatus = "normal" | "full";
-
-export type Baker = {
-  address: string;
-  name: string;
-  logoURL: string;
-  nominalYield: `${number} %`;
-  capacityStatus: CapacityStatus;
-};
-
-// type used by UI to facilitate business logic of current delegation data
-export type Delegation = {
-  // delegator address
-  address: string;
-  // if not defined, we need to render "Unknown" on the UI. we don't know who is delegator.
-  baker: Baker | null | undefined;
-  // operation related to delegation (to know the date info)
-  operation: Operation;
-  // true if the delegation is pending (optimistic update)
-  isPending: boolean;
-  // true if a receive should inform it will top up the delegation
-  receiveShouldWarnDelegation: boolean;
-  // true if a send should inform it will top down the delegation
-  sendShouldWarnDelegation: boolean;
-};
-
-type API_BAKER = {
-  address: string;
-  name: string;
-  logo: string;
-  balance: number;
-  stakingBalance: number;
-  stakingCapacity: number;
-  maxStakingBalance: number;
-  freeSpace: number;
-  fee: number;
-  minDelegation: number;
-  payoutDelay: number;
-  payoutPeriod: number;
-  openForDelegation: true;
-  estimatedRoi: number;
-  serviceType: string;
-  serviceHealth: string;
-  payoutTiming: string;
-  payoutAccuracy: string;
-  audit: string;
-  insuranceCoveragenumber: number;
-};
+import { API_BAKER, Baker, Delegation } from "./types";
 
 const cache = makeLRUCache(
   async (): Promise<Baker[]> => {
